Network.TypedProtocol.Stateful.Codec

Description

Stateful codec. This module is intended to be imported qualified.

Defining and using Codecs

data Codec ps failure (f :: ps -> Type) m bytes

A stateful codec.

Constructors

Codec

Fields

encode :: forall (st :: ps) (st' :: ps). StateTokenI st => ActiveState st => f st -> Message ps st st' -> bytes

decode :: forall (st :: ps). ActiveState st => StateToken st -> f st -> m (DecodeStep bytes failure m (SomeMessage st))

hoistCodec :: Functor n => (forall x. m x -> n x) -> Codec ps failure f m bytes -> Codec ps failure f n bytes

isoCodec :: Functor m => (bytes -> bytes') -> (bytes' -> bytes) -> Codec ps failure f m bytes -> Codec ps failure f m bytes'

mapFailureCodec :: Functor m => (failure -> failure') -> Codec ps failure f m bytes -> Codec ps failure' f m bytes

liftCodec :: Codec ps failure m bytes -> Codec ps failure f m bytes

Incremental decoding

data DecodeStep bytes failure (m :: Type -> Type) a

Constructors

DecodePartial (Maybe bytes -> m (DecodeStep bytes failure m a))

DecodeDone a (Maybe bytes)

DecodeFail failure

runDecoder :: Monad m => [bytes] -> DecodeStep bytes failure m a -> m (Either failure a)

runDecoderPure :: Monad m => (forall b. m b -> b) -> m (DecodeStep bytes failure m a) -> [bytes] -> Either failure a

Related types

SomeMessage

data SomeMessage (st :: ps) where

Constructors

SomeMessage :: forall ps (st :: ps) (st' :: ps). (StateTokenI st, StateTokenI st', ActiveState st) => Message ps st st' -> SomeMessage st

StateToken

type family StateToken :: ps -> Type

class StateTokenI (st :: ps) where

Methods

stateToken :: StateToken st

ActiveState

type ActiveState (st :: ps) = IsActiveState st (StateAgency st)

PeerRole

data PeerRole

Constructors

AsClient

AsServer

CodecFailure

data CodecFailure

Constructors

CodecFailureOutOfInput

CodecFailure String

Testing codec properties

data AnyMessage ps (f :: ps -> Type) where

Any message for a protocol, with a StateTokenI constraint which gives access to protocol state.

Used where we don't know statically what the state type is, but need the agency and message to match each other.

Constructors

AnyMessage

Fields

:: forall ps f (st :: ps) (st' :: ps). (StateTokenI st, ActiveState st)

=> f st

local state

-> Message ps (st :: ps) (st' :: ps)

protocol messsage

-> AnyMessage ps f

Bundled Patterns

pattern AnyMessageAndAgency :: forall ps f. () => forall (st :: ps) (st' :: ps). (StateTokenI st, ActiveState st) => StateToken st -> f st -> Message ps st st' -> AnyMessage ps f
A convenient pattern synonym which unwrap AnyMessage giving both the singleton for the state and the message.

showAnyMessage :: forall ps f. (forall st st'. Show (Message ps st st'), forall st. Show (f st)) => AnyMessage ps f -> String

showAnyMessage is can be used to provide Show instance for AnyMessage if showing Message is independent of the state or one accepts showing only partial information included in message constructors or accepts message constructors to carry Show instances for its arguments. Note that the proper solution is to define a custom `Show (AnyMessage ps f)` instance for a protocol ps, which give access to the state functor f in scope of show.

prop_codecM :: forall ps failure f m bytes. (Monad m, Eq (AnyMessage ps f)) => Codec ps failure f m bytes -> AnyMessage ps f -> m Bool

The Codec round-trip property: decode after encode gives the same message. Every codec must satisfy this property.

prop_codec :: forall ps failure f m bytes. (Monad m, Eq (AnyMessage ps f)) => (forall a. m a -> a) -> Codec ps failure f m bytes -> AnyMessage ps f -> Bool

The Codec round-trip property in a pure monad.

prop_codec_splitsM

Arguments

:: forall ps failure f m bytes. (Monad m, Eq (AnyMessage ps f))

=> (bytes -> [[bytes]])
alternative re-chunkings of serialised form

-> Codec ps failure f m bytes

-> AnyMessage ps f

-> m Bool

A variant on the codec round-trip property: given the encoding of a message, check that decode always gives the same result irrespective of how the chunks of input are fed to the incremental decoder.

This property guards against boundary errors in incremental decoders. It is not necessary to check this for every message type, just for each generic codec construction. For example given some binary serialisation library one would write a generic adaptor to the codec interface. This adaptor has to deal with the incremental decoding and this is what needs to be checked.

prop_codec_splits :: forall ps failure f m bytes. (Monad m, Eq (AnyMessage ps f)) => (bytes -> [[bytes]]) -> (forall a. m a -> a) -> Codec ps failure f m bytes -> AnyMessage ps f -> Bool

Like prop_codec_splitsM but run in a pure monad m, e.g. Identity.

prop_codecs_compatM :: forall ps failure f m bytes. (Monad m, Eq (AnyMessage ps f), forall a. Monoid a => Monoid (m a)) => Codec ps failure f m bytes -> Codec ps failure f m bytes -> AnyMessage ps f -> m Bool

Compatibility between two codecs of the same protocol. Encode a message with one codec and decode it with the other one, then compare if the result is the same as initial message.

prop_codecs_compat :: forall ps failure f m bytes. (Monad m, Eq (AnyMessage ps f), forall a. Monoid a => Monoid (m a)) => (forall a. m a -> a) -> Codec ps failure f m bytes -> Codec ps failure f m bytes -> AnyMessage ps f -> Bool

Like prop_codecs_compatM but run in a pure monad m, e.g. Identity.