Network.TypedProtocol.Stateful.Driver

Description

Actions for running Peers with a Driver. This module should be imported qualified.

DriverIngerface

data Driver ps (pr :: PeerRole) bytes failure dstate f m

Constructors

Driver

Fields

sendMessage :: forall (st :: ps) (st' :: ps). StateTokenI st => StateTokenI st' => ActiveState st => ReflRelativeAgency (StateAgency st) WeHaveAgency (Relative pr (StateAgency st)) -> f st -> Message ps st st' -> m ()

Send a message.

recvMessage :: forall (st :: ps). StateTokenI st => ActiveState st => ReflRelativeAgency (StateAgency st) TheyHaveAgency (Relative pr (StateAgency st)) -> f st -> dstate -> m (SomeMessage st, dstate)

Receive a message, a blocking action which reads from the network and runs the incremental decoder until a full message is decoded.

initialDState :: dstate

Initial decoder state.

Running a peer

runPeerWithDriver :: forall ps (st :: ps) pr bytes failure dstate (f :: ps -> Type) m a. MonadSTM m => Driver ps pr bytes failure dstate f m -> f st -> Peer ps pr st f m a -> m (a, dstate)

Run a peer with the given driver.

This runs the peer to completion (if the protocol allows for termination).

NOTE: this function threads local state (i.e. f) through evolution of a protocol (i.e. Peer).

Re-exports

data SomeMessage (st :: ps) where

Constructors

SomeMessage :: forall ps (st :: ps) (st' :: ps). (StateTokenI st, StateTokenI st', ActiveState st) => Message ps st st' -> SomeMessage st

data DecodeStep bytes failure (m :: Type -> Type) a

Constructors

DecodePartial (Maybe bytes -> m (DecodeStep bytes failure m a))

DecodeDone a (Maybe bytes)

DecodeFail failure