Network.TypedProtocol.Stateful.Proofs

Description

Proofs about the typed protocol framework.

It also provides helpful testing utilities.

Documentation

connect :: forall ps (pr :: PeerRole) (st :: ps) (f :: ps -> Type) m a b. (MonadSTM m, SingI pr) => f st -> Peer ps pr st f m a -> Peer ps (FlipAgency pr) st f m b -> m (a, b, TerminalStates ps)

data TerminalStates ps where

Constructors

TerminalStates :: forall ps (st :: ps). StateAgency st ~ 'NobodyAgency => StateToken st -> StateToken st -> TerminalStates ps

removeState :: Functor m => f st -> Peer ps pr st f m a -> Peer ps pr NonPipelined st m a

Remove state for non-pipelined peers.

TODO: There's a difficulty to write removeState for pipelined peers which is type safe. The Peer doesn't track all pipelined transitions, just the depth of pipelining, so we cannot push `f st` to a queue which type is linked to Peer. For a similar reason there's no way to write forgetPipelined function.

However, this is possible if Peer tracks all transitions.